Consolidate approach prompts into a single prompts.py module
Move all prompt text for the agentic, CDL, and genplan approaches into a new
source-free robocode/prompts.py with composition helpers (build_system_prompt,
build_interface_spec, build_mcp_tool_lines, build_agentic_prompt,
build_cdl_prompt). Blackbox is now a named delta selected by a bool rather than
parallel full copies, and the agentic/CDL system prompts are composed from
shared fragments (file discipline, subagents, blackbox-discovery, learning
clause) so the silent drift that had accumulated cannot reappear.

Two fragments that had drifted are unified to one canonical version:
- the blackbox interaction spec (CDL's make_primitives paragraph now matches the
  monolithic approach's runnable-code-block form);
- the "read source ... write an optimal approach class" learning clause, now
  shared by both intros (previously present only on the agentic intro).

genplan's prompt constants are relocated verbatim, kept textually faithful to
the upstream llm-genplan method. The refactor is behavior-preserving for the
agentic approach; the CDL system prompt and blackbox task prompt change as noted
above, verified with a before/after byte-equality snapshot. Adds
tests/test_prompts.py.
